SQLite versions of topK and topKWithIndex
